requirements.txt
torch>=1.9.0
transformers>=4.0.0
pandas>=1.3.0
imap-tools>=0.54.0
python-dotenv>=0.19.0